fix for long emms
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@11073 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ globals.h
@@ -802,7 +802,7 @@ struct s_emmcache
 {
 	uchar			emmd5[CS_EMMSTORESIZE];
 	uchar			type;
-	uchar			len;
+	uint16_t		len;
 	uchar			emm[MAX_EMM_SIZE];
 	struct timeb    firstseen;
 	struct timeb    lastseen;
